Fjernet TPS-trivia #deploy-app-tilgang-analyse-service
