--- apps/testnorge-statisk-data-forvalter/src/main/java/no/nav/registre/sdforvalter/consumer/rs/tpsf/request/SendToTpsRequest.java
@@ -1,18 +1,0 @@
-package no.nav.registre.sdforvalter.consumer.rs.tpsf.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SendToTpsRequest {
-
-    private String environment;
-    private List<Long> ids;
-}
