--- apps/testnorge-statisk-data-forvalter/src/main/java/no/nav/registre/sdforvalter/database/model/TpsIdentTagModel.java
@@ -1,41 +1,0 @@
-package no.nav.registre.sdforvalter.database.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "TPS_IDENTER_TAGS")
-@Builder
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@AllArgsConstructor
-@NoArgsConstructor
-public class TpsIdentTagModel extends AuditModel {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "ID")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "IDENT")
-    private TpsIdentModel ident;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "TAG")
-    private TagModel tag;
-}
