--- apps/testnorge-statisk-data-forvalter/src/main/java/no/nav/registre/sdforvalter/domain/TpsIdentListe.java
@@ -1,14 +1,0 @@
-package no.nav.registre.sdforvalter.domain;
-
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@NoArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-public final class TpsIdentListe extends FasteDataListe<TpsIdent> {
-    public TpsIdentListe(List<TpsIdent> liste) {
-        super(liste);
-    }
-}
