+++ apps/testnorge-statisk-data-forvalter/src/main/java/no/nav/registre/sdforvalter/domain/person/Person.java
@@ -3,9 +3,6 @@ package no.nav.registre.sdforvalter.domain.person;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.NoArgsConstructor;
 import lombok.Value;
-
-import no.nav.registre.sdforvalter.domain.TpsIdent;
-import no.nav.testnav.libs.dto.person.v1.AdresseDTO;
 import no.nav.testnav.libs.dto.person.v1.PersonDTO;
 
 
@@ -25,15 +22,6 @@ public class Person {
     @JsonProperty("by")
     String city;
 
-    public Person(TpsIdent ident) {
-        fnr = ident.getFnr();
-        firstName = ident.getFirstName();
-        lastName = ident.getLastName();
-        address = ident.getAddress();
-        postNr = ident.getPostNr();
-        city = ident.getCity();
-    }
-
     public Person(PersonDTO dto) {
         fnr = dto.getIdent();
         firstName = dto.getFornavn();
@@ -41,14 +29,5 @@ public class Person {
         address = dto.getAdresse() != null ? dto.getAdresse().getGatenavn() : null;
         postNr = dto.getAdresse() != null ? dto.getAdresse().getPostnummer() : null;
         city = dto.getAdresse() != null ? dto.getAdresse().getPoststed() : null;
-    }
-
-    public PersonDTO toDTO () {
-        return PersonDTO.builder()
-                .ident(fnr)
-                .fornavn(firstName)
-                .etternavn(lastName)
-                .adresse(address != null ? new AdresseDTO(address, postNr, city, null) : null)
-                .build();
     }
 }
