+++ apps/testnorge-statisk-data-forvalter/src/main/java/no/nav/registre/sdforvalter/provider/rs/FileController.java
@@ -1,18 +1,13 @@
 package no.nav.registre.sdforvalter.provider.rs;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import no.nav.registre.sdforvalter.adapter.EregAdapter;
-import no.nav.registre.sdforvalter.adapter.TpsIdenterAdapter;
 import no.nav.registre.sdforvalter.converter.csv.EregCsvConverter;
-import no.nav.registre.sdforvalter.converter.csv.TpsIdentCsvConverter;
 import no.nav.registre.sdforvalter.domain.Ereg;
 import no.nav.registre.sdforvalter.domain.EregListe;
-import no.nav.registre.sdforvalter.domain.TpsIdent;
-import no.nav.registre.sdforvalter.domain.TpsIdentListe;
-import no.nav.registre.sdforvalter.service.IdentService;
 import no.nav.testnav.libs.dto.organisasjonfastedataservice.v1.Gruppe;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -24,7 +19,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import jakarta.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -46,15 +40,13 @@ import java.util.stream.Collectors;
 public class FileController {
 
     private final EregAdapter eregAdapter;
-    private final TpsIdenterAdapter tpsIdenterAdapter;
-    private final IdentService identService;
 
     @Operation(summary = "Hent organisasjoner fra Team Dollys database")
     @GetMapping("/ereg")
     public void exportEreg(@RequestParam(name = "gruppe", required = false) Gruppe gruppe, HttpServletResponse response) throws IOException {
         response.setContentType("text/csv");
         response.setCharacterEncoding(StandardCharsets.UTF_8.toString());
-        response.setHeader("Content-Disposition", "attachment; filename=ereg-" + LocalDateTime.now().toString() + ".csv");
+        response.setHeader("Content-Disposition", "attachment; filename=ereg-" + LocalDateTime.now() + ".csv");
         EregListe eregListe = eregAdapter.fetchBy(gruppe != null ? gruppe.name() : null);
 
         EregCsvConverter.inst().write(response.getWriter(), eregListe.getListe());
@@ -65,28 +57,6 @@ public class FileController {
     public ResponseEntity<HttpStatus> importEreg(@RequestParam("file") MultipartFile file) throws IOException {
         List<Ereg> list = EregCsvConverter.inst().read(new InputStreamReader(file.getInputStream(), StandardCharsets.UTF_8));
         eregAdapter.save(new EregListe(list));
-        return ResponseEntity.ok().build();
-    }
-
-    @Operation(summary = "Hent personer fra Team Dollys database")
-    @GetMapping("/tpsIdenter")
-    public void exportTpsIdenter(@RequestParam(name = "gruppe", required = false) Gruppe gruppe, HttpServletResponse response) throws IOException {
-        response.setContentType("text/csv");
-        response.setCharacterEncoding(StandardCharsets.UTF_8.toString());
-        response.setHeader("Content-Disposition", "attachment; filename=tpsIdent-" + LocalDateTime.now().toString() + ".csv");
-
-        TpsIdentListe tpsIdentListe = tpsIdenterAdapter.fetchBy(gruppe.name());
-        TpsIdentCsvConverter.inst().write(response.getWriter(), tpsIdentListe.getListe());
-    }
-
-    @Operation(summary = "Legg til personer i Team Dollys database")
-    @PostMapping(path = "/tpsIdenter", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<?> importTpsIdenter(@RequestParam("file") MultipartFile file,
-                                              @Parameter(description = "Hvis true settes tilfeldig navn på personer uten fornavn og etternavn")
-                                              @RequestParam(name = "Generer manglende navn", defaultValue = "false") Boolean genererManglendeNavn) throws IOException {
-        List<TpsIdent> list = TpsIdentCsvConverter.inst().read(cleanInput(file.getInputStream()));
-
-        identService.save(new TpsIdentListe(list), genererManglendeNavn);
         return ResponseEntity.ok().build();
     }
