+++ apps/testnorge-statisk-data-forvalter/src/test/java/no/nav/registre/sdforvalter/provider/rs/FileControllerIntegrationTest.java
@@ -2,12 +2,12 @@ package no.nav.registre.sdforvalter.provider.rs;
 
 import no.nav.registre.sdforvalter.database.model.GruppeModel;
 import no.nav.registre.sdforvalter.database.model.OpprinnelseModel;
-import no.nav.registre.sdforvalter.database.model.TpsIdentModel;
-import no.nav.registre.sdforvalter.database.repository.*;
-import no.nav.registre.sdforvalter.domain.TpsIdent;
+import no.nav.registre.sdforvalter.database.repository.EregTagRepository;
+import no.nav.registre.sdforvalter.database.repository.GruppeRepository;
+import no.nav.registre.sdforvalter.database.repository.OpprinnelseRepository;
+import no.nav.registre.sdforvalter.database.repository.TagRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,16 +16,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-
-import java.util.List;
-import java.util.Set;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.reset;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.jwt;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @ExtendWith(MockitoExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -40,8 +32,6 @@ class FileControllerIntegrationTest {
     private MockMvc mvc;
 
     @Autowired
-    private TpsIdenterRepository tpsIdenterRepository;
-    @Autowired
     private GruppeRepository gruppeRepository;
     @Autowired
     private OpprinnelseRepository opprinnelseRepository;
@@ -49,8 +39,6 @@ class FileControllerIntegrationTest {
     private TagRepository tagRepository;
     @Autowired
     private EregTagRepository eregTagRepository;
-    @Autowired
-    private TpsIdentTagRepository tpsIdentTagRepository;
 
     @BeforeEach
     public void setup() {
@@ -61,87 +49,12 @@ class FileControllerIntegrationTest {
         opprinnelseRepository.save(opprinnelseModel);
     }
 
-    public void assertListOfPersonsFromCsvIsSavedInDatabase(List<TpsIdentModel> expectedTpsIdenterListe, String csvInnhold) throws Exception {
-
-        MockHttpServletRequestBuilder builder = MockMvcRequestBuilders
-                .multipart("/api/v1/faste-data/file/tpsIdenter")
-                .file("file", csvInnhold.getBytes())
-                .with(jwt())
-                .characterEncoding("UTF_8");
-
-        mvc.perform(builder).andExpect(status().isOk());
-
-        assertThat(tpsIdenterRepository.findAll())
-                .containsAll(expectedTpsIdenterListe);
-
-    }
-
     @AfterEach
     public void cleanUp() {
         reset();
         eregTagRepository.deleteAll();
-        tpsIdentTagRepository.deleteAll();
         tagRepository.deleteAll();
-        tpsIdenterRepository.deleteAll();
         opprinnelseRepository.deleteAll();
         gruppeRepository.deleteAll();
-    }
-
-    @Test
-    void should_save_one_ident_from_csv_to_tps_ident_database() throws Exception {
-        String csvInnhold = "FNR*;Fornavn;Etternavn;Adresse;Postnummer;Poststed;Gruppe;Opprinnelse;Tags\n" +
-                "12345678912;Dolly;Dollesen;Dollygata 2;9999;Dollyville;Gruppen;Test;OTP";
-
-        TpsIdent expectedTpsPerson = TpsIdent.builder()
-                .fnr("12345678912")
-                .firstName("Dolly")
-                .lastName("Dollesen")
-                .address("Dollygata 2")
-                .postNr("9999")
-                .city("Dollyville")
-                .tags(Set.of("OTP"))
-                .build();
-        List<TpsIdentModel> expectedTpsIdenterListe = List.of(new TpsIdentModel(
-                        expectedTpsPerson,
-                        new OpprinnelseModel(null, "Test"),
-                        new GruppeModel(null, "Gruppen", "Gruppenbeskrivelse")
-                )
-        );
-
-        assertListOfPersonsFromCsvIsSavedInDatabase(expectedTpsIdenterListe, csvInnhold);
-    }
-
-    @Test
-    void should_save_two_idents_from_csv_to_tps_ident_database() throws Exception {
-        String csvInnhold = """
-                FNR*;Fornavn;Etternavn;Adresse;Postnummer;Poststed;Gruppe;Opprinnelse;Tags
-                12345678910;Dolly;Dollesen;Dollygata 2;9999;Dollyville;Gruppen;Test;OTP
-                12345678911;Donald;Dollesen;Dollygata 3;2222;Dollyby;Gruppen;Test;""";
-
-        TpsIdent expectedTpsPerson1 = TpsIdent.builder()
-                .fnr("12345678910")
-                .firstName("Dolly")
-                .lastName("Dollesen")
-                .address("Dollygata 2")
-                .postNr("9999")
-                .city("Dollyville")
-                .tags(Set.of("OTP"))
-                .build();
-        TpsIdent expectedTpsPerson2 = TpsIdent.builder()
-                .fnr("12345678911")
-                .firstName("Donald")
-                .lastName("Dollesen")
-                .address("Dollygata 3")
-                .postNr("2222")
-                .city("Dollyby")
-                .build();
-        var opprinnelse = new OpprinnelseModel(null, "Test");
-        var gruppe = new GruppeModel(null, "Gruppen", "Gruppenbeskrivelse");
-        List<TpsIdentModel> expectedTpsIdenterListe = List.of(
-                new TpsIdentModel(expectedTpsPerson1, opprinnelse, gruppe),
-                new TpsIdentModel(expectedTpsPerson2, opprinnelse, gruppe)
-        );
-
-        assertListOfPersonsFromCsvIsSavedInDatabase(expectedTpsIdenterListe, csvInnhold);
     }
 }
